requirements.txt
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
langchain==0.1.11
langchain-openai==0.0.5
openai==1.14.0
pandas==2.2.0
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3 